1st commit

+++ resources/views/contact/contact.blade.php
@@ -1,11 +1,14 @@
 @extends('navbar')
 @section('navbar')
 
+    @inject('privilege', 'App\Http\Controllers\check_privileges')
+
 <div class="row">
 <div class="col-sm-1"></div>
 <div class="col-sm-10">
 
-<form id="form2" method="post" action="<?php $_SERVER['PHP_SELF']?>?loc=contact&action=view&alert=">
+<form id="form2" method="post" action="contact/search">
+    <input type="hidden" id="_token" name="_token" value="{{ csrf_token() }}">
 
 <fieldset class="scheduler-border">
 <legend class="scheduler-border">
@@ -68,11 +71,11 @@
     <div class="col-sm-1"></div>
     <div class="col-sm-2">
 
-        <?php // if ($privilege->add_privilege("contacts") == "1"){    //check privilege
+        <?php  if ($privilege->add_privilege("contacts") == "1"){    //check privilege
         ?>
-        <a href="<?php $_SERVER['PHP_SELF']?>?loc=contact&action=add&alert=">Add Contact</a>
-        <?php  //
-        //}
+        <a href="{{ 'contact/create' }}">Add Contact</a>
+        <?php
+        }
             ?>
 
     </div>
@@ -102,7 +105,7 @@
             foreach ($result as $row)
             {
 
-            //if($privilege->view_prvilege($row['group'],"contacts",$row['assignedto']) == "1" ){   // check user can viwe this data
+            if($privilege->view_prvilege($row->group_id,"contacts",$row->assignedto) == "1" ){   // check user can viwe this data
 
             ?>
             <tr>
@@ -111,7 +114,7 @@
 
                     if($b=="contact_owner"){
                         echo "<td><a1>".$row->user_name."</a1></td>";
-                    }elseif ($b=="group") {
+                    }elseif ($b=="group_id") {
                         echo "<td><a1>".$row->group_name."</a1></td>";
                     }elseif ($b=="assignedto") {
                         echo "<td><a1>".$row->Assign."</a1></td>";
@@ -124,14 +127,14 @@
                     }
                 }
                 ?>
-                <td><?php // if ($privilege->edit_privilege($row['group'],"contacts",$row['assignedto']) == "1" ){ ?><a href="<?php $_SERVER['PHP_SELF']?>?loc=contact&action=edit&cid=<?php echo $row->id; ?>" class="glyphicon glyphicon-pencil" ></a><?php //} ?>
-                    <?php // if ($privilege->delete_privilege($row['group'],"contacts",$row['assignedto']) == "1" ){ ?><a href="#" onclick="AlertIt('functions/delete.php?table=contacts&id=<?php echo $row->id; ?>','1');" class="glyphicon glyphicon-remove-circle" style="color:red" ></a> <?php // } ?>
+                <td><?php  if ($privilege->edit_privilege($row->group_id,"contacts",$row->assignedto) == "1" ){ ?><a href="contact/{{$row->id}}/edit" class="glyphicon glyphicon-pencil" ></a><?php } ?>
+                    <?php  if ($privilege->delete_privilege($row->group_id,"contacts",$row->assignedto ) == "1" ){ ?><a href="#" onclick="deleted({{ $row->id }})" class="glyphicon glyphicon-remove-circle" style="color:red" ></a> <?php  } ?>
                 </td>
             </tr>
 
             <?php
 
-           // }//end of if
+            }//end of if
             }// end of while
 
             ?>
@@ -156,6 +159,43 @@
 
 </body>
 </html>
+
+<script type="text/javascript">
+
+    var root_url = "<?php echo Request::root(); ?>/"; // put this in php file
+
+    function deleted(id){
+
+        var token   = "{{ csrf_token() }}";
+        var method  = "DELETE";
+
+        var r = confirm("Confirm Delete!");
+
+        if (r == true) {
+
+            $.ajax({
+                type: 'post',
+                url: root_url + 'contact/' + id,
+                data: {"_token": token, "_method": method},
+
+                success: function () {
+
+                    alert('Delete Successful');
+                    location.reload();
+
+                },
+                error: function (xhr, textStatus, error) {
+                    alert(error);
+                }
+            });
+
+        }
+    }
+
+    function load_search(){
+        $("#search").load(root_url+"contact_ad_search");
+    }
+</script>
 
     @stop
